+++ packages/web-components/src/search-widget/types/index.d.ts
@@ -1,2 +1,2 @@
-export * from './PickedPropertyModel'
-export * from './PickedPropertyImageModel'
+export * from './picked-properties'
+export * from './picked-property-image'
