+++ packages/web-components/src/search-widget/types/picked-properties.d.ts
