+++ packages/web-components/src/search-widget/types/picked-property-image.d.ts
